+++ src/NPoco.Abstractions/Linq/IAsyncQueryProvider.cs
@@ -23,13 +23,13 @@ namespace NPoco.Linq
         Task<T?> SingleOrDefault(Expression<Func<T, bool>> whereExpression, CancellationToken cancellationToken = default);
         Task<T> Single(CancellationToken cancellationToken = default);
         Task<T> Single(Expression<Func<T, bool>> whereExpression, CancellationToken cancellationToken = default);
-        Task<int> Count(CancellationToken cancellationToken = default);
-        Task<int> Count(Expression<Func<T, bool>> whereExpression, CancellationToken cancellationToken = default);
+        Task<long> Count(CancellationToken cancellationToken = default);
+        Task<long> Count(Expression<Func<T, bool>> whereExpression, CancellationToken cancellationToken = default);
         Task<bool> Any(CancellationToken cancellationToken = default);
         Task<bool> Any(Expression<Func<T, bool>> whereExpression, CancellationToken cancellationToken = default);
-        Task<Page<T>> ToPage(int page, int pageSize, CancellationToken cancellationToken = default);
+        Task<Page<T>> ToPage(long page, long pageSize, CancellationToken cancellationToken = default);
         Task<List<T2>> ProjectTo<T2>(Expression<Func<T, T2>> projectionExpression, CancellationToken cancellationToken = default);
-        Task<Page<T2>> ToProjectedPage<T2>(Expression<Func<T, T2>> projectionExpression, int page, int pageSize, CancellationToken cancellationToken = default);
+        Task<Page<T2>> ToProjectedPage<T2>(Expression<Func<T, T2>> projectionExpression, long page, long pageSize, CancellationToken cancellationToken = default);
         Task<List<T2>> Distinct<T2>(Expression<Func<T, T2>> projectionExpression, CancellationToken cancellationToken = default);
         Task<List<T>> Distinct(CancellationToken cancellationToken = default);
     }
@@ -44,8 +44,8 @@ namespace NPoco.Linq
         T? SingleOrDefault(Expression<Func<T, bool>> whereExpression);
         T Single();
         T Single(Expression<Func<T, bool>> whereExpression);
-        int Count();
-        int Count(Expression<Func<T, bool>> whereExpression);
+        long Count();
+        long Count(Expression<Func<T, bool>> whereExpression);
         bool Any();
         bool Any(Expression<Func<T, bool>> whereExpression);
         List<T> ToList();
@@ -53,9 +53,9 @@ namespace NPoco.Linq
         IEnumerable<T> ToEnumerable();
         List<dynamic> ToDynamicList();
         IEnumerable<dynamic> ToDynamicEnumerable();
-        Page<T> ToPage(int page, int pageSize);
+        Page<T> ToPage(long page, long pageSize);
         List<T2> ProjectTo<T2>(Expression<Func<T, T2>> projectionExpression);
-        Page<T2> ToProjectedPage<T2>(Expression<Func<T, T2>> projectionExpression, int page, int pageSize);
+        Page<T2> ToProjectedPage<T2>(Expression<Func<T, T2>> projectionExpression, long page, long pageSize);
         List<T2> Distinct<T2>(Expression<Func<T, T2>> projectionExpression);
         List<T> Distinct();
         Task<List<T>> ToListAsync(CancellationToken cancellationToken = default);
@@ -69,13 +69,13 @@ namespace NPoco.Linq
         Task<T?> SingleOrDefaultAsync(Expression<Func<T, bool>> whereExpression, CancellationToken cancellationToken = default);
         Task<T> SingleAsync(CancellationToken cancellationToken = default);
         Task<T> SingleAsync(Expression<Func<T, bool>> whereExpression, CancellationToken cancellationToken = default);
-        Task<int> CountAsync(CancellationToken cancellationToken = default);
-        Task<int> CountAsync(Expression<Func<T, bool>> whereExpression, CancellationToken cancellationToken = default);
+        Task<long> CountAsync(CancellationToken cancellationToken = default);
+        Task<long> CountAsync(Expression<Func<T, bool>> whereExpression, CancellationToken cancellationToken = default);
         Task<bool> AnyAsync(CancellationToken cancellationToken = default);
         Task<bool> AnyAsync(Expression<Func<T, bool>> whereExpression, CancellationToken cancellationToken = default);
-        Task<Page<T>> ToPageAsync(int page, int pageSize, CancellationToken cancellationToken = default);
+        Task<Page<T>> ToPageAsync(long page, long pageSize, CancellationToken cancellationToken = default);
         Task<List<T2>> ProjectToAsync<T2>(Expression<Func<T, T2>> projectionExpression, CancellationToken cancellationToken = default);
-        Task<Page<T2>> ToProjectedPageAsync<T2>(Expression<Func<T, T2>> projectionExpression, int page, int pageSize, CancellationToken cancellationToken = default);
+        Task<Page<T2>> ToProjectedPageAsync<T2>(Expression<Func<T, T2>> projectionExpression, long page, long pageSize, CancellationToken cancellationToken = default);
         Task<List<T2>> DistinctAsync<T2>(Expression<Func<T, T2>> projectionExpression, CancellationToken cancellationToken = default);
         Task<List<T>> DistinctAsync(CancellationToken cancellationToken = default);
     }
